decompose webserver directory
