+++ frontend/vite.config.js
@@ -1,14 +1,12 @@
 import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
-import fs from 'fs';
 
 export default defineConfig({
   plugins: [vue()],
-  root: 'frontend', 
+  root: './', 
   build: {
-    outDir: '../dist',  
+    outDir: './dist',  
     emptyOutDir: true,  
   },
-  base: "./", 
-
+  base: "/website/", // base path for the website
 });
