--- webserver/dockerfile
@@ -1,35 +1,0 @@
-FROM node:18
-
-# Install dependencies for AWS CLI v2 installation
-RUN apt-get update && apt-get install -y \
-    curl \
-    unzip \
-    && rm -rf /var/lib/apt/lists/*
-
-# Install AWS CLI v2
-RUN curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip" \
-    && unzip awscliv2.zip \
-    && ./aws/install \
-    && rm -rf awscliv2.zip aws/
-
-# Set AWS access
-ARG AWS_ACCESS_KEY_ID
-ARG AWS_SECRET_ACCESS_KEY
-ENV AWS_ACCESS_KEY_ID=$AWS_ACCESS_KEY_ID
-ENV AWS_SECRET_ACCESS_KEY=$AWS_SECRET_ACCESS_KEY
-ENV AWS_REGION=eu-north-1
-
-WORKDIR /app
-
-# Install dependencies for the webapp
-COPY . .
-RUN npm install
-
-# Build webapp
-RUN npm run build
-
-# Open port 3000 for the webserver
-EXPOSE 3000
-
-# start the webserver
-CMD ["npm","run","start"]
