--- webserver/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "document-stream-filter",
-  "version": "1.0.0",
-  "type": "module",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "vite",
-    "build": "vite build && aws s3 sync dist s3://document-stream-filter-bucket --delete",
-    "start": "node server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/eliasinguanta/document-stream-filter.git"
-  },
-  "keywords": [],
-  "author": "Elias Inguanta",
-  "license": "MIT",
-  "homepage": "https://github.com/eliasinguanta/document-stream-filter#readme",
-  "dependencies": {
-    "@aws-sdk/client-dynamodb": "^3.782.0",
-    "@aws-sdk/client-s3": "^3.782.0",
-    "@aws-sdk/lib-dynamodb": "^3.778.0",
-    "@aws-sdk/util-dynamodb": "^3.782.0",
-    "aws-sdk": "^2.1692.0",
-    "bootstrap": "^5.3.4",
-    "cors": "^2.8.5",
-    "dotenv": "^16.4.7",
-    "express": "^4.21.2",
-    "formidable": "^3.5.2",
-    "multer": "^1.4.5-lts.2",
-    "node-fetch": "^3.3.2",
-    "http-proxy-middleware": "^3.0.5"
-  },
-  "devDependencies": {
-    "@vitejs/plugin-vue": "^5.2.3",
-    "vite": "^6.2.5",
-    "vue": "^3.5.13"
-  }
-}
