+++ website/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "description": "send the homepage (html, css, client-side js)",
   "main": "server.js",
+  "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node server.js"
@@ -10,7 +11,7 @@
   "author": "Elias Inguanta",
   "license": "MIT",
   "dependencies": {
-    "@aws-sdk/client-s3": "^3.800.0",
+    "@aws-sdk/client-s3": "^3.802.0",
     "express": "^5.1.0"
   }
 }
