+++ website/server.js
@@ -1,11 +1,10 @@
 import express from "express";
 
-import {getWebsiteFromS3} from "./backend/s3_api.js";
+import {getWebsiteFromS3} from "./s3.js";
 
 const app = express();
 const PORT = 80;
 const PREAFIX = "/website";
-
 // health check
 // basically a ping to check if the server is running
 // For AWS ELB a health check endpoint is required (but i also just could use the "/" endpoint)
@@ -14,11 +13,9 @@ app.get(PREAFIX+"/health", (req, res) => {
 });
 
 // get the website
-app.get(PREAFIX, getWebsiteFromS3, async (req, res) => {
-  res.end();
-});
+app.use(PREAFIX, getWebsiteFromS3);
 
 //start http server
 app.listen(PORT, () => {
-  console.log(`server runs on http://localhost:${PORT}`);
+  console.log(`server runs on http://localhost:${PORT}${PREAFIX}`);
 });
